+++ NEWS.md
@@ -6,6 +6,7 @@
 - add digits for mass accuracy
 - fix the ms1 larger issue in getpmd
 - add support for quantitative paired peaks list selection for specific reaction
+- add support for target pmd and compound analysis for reaction chain
 
 # pmd 0.1.2